requirements.txt
alembic==1.14.1
azure-core==1.32.0
azure-storage-blob==12.24.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
gunicorn==23.0.0
pyodbc==5.2.0
azure-storage-blob==12.24.1
azure-core==1.32.0
alembic==1.14.1